[HW12] Organized cmd and configs using cobra & viper
