--- hw12_13_14_15_calendar/cmd/calendar/config.go
@@ -1,20 +1,0 @@
-package main
-
-// При желании конфигурацию можно вынести в internal/config.
-// Организация конфига в main принуждает нас сужать API компонентов, использовать
-// при их конструировании только необходимые параметры, а также уменьшает вероятность циклической зависимости.
-type Config struct {
-	Logger LoggerConf
-	// TODO
-}
-
-type LoggerConf struct {
-	Level string
-	// TODO
-}
-
-func NewConfig() Config {
-	return Config{}
-}
-
-// TODO
