--- hw12_13_14_15_calendar/cmd/calendar/version.go
@@ -1,27 +1,0 @@
-package main
-
-import (
-	"encoding/json"
-	"fmt"
-	"os"
-)
-
-var (
-	release   = "UNKNOWN"
-	buildDate = "UNKNOWN"
-	gitHash   = "UNKNOWN"
-)
-
-func printVersion() {
-	if err := json.NewEncoder(os.Stdout).Encode(struct {
-		Release   string
-		BuildDate string
-		GitHash   string
-	}{
-		Release:   release,
-		BuildDate: buildDate,
-		GitHash:   gitHash,
-	}); err != nil {
-		fmt.Printf("error while decode version info: %v\n", err)
-	}
-}
