+++ hw12_13_14_15_calendar/go.mod
@@ -1,3 +1,8 @@
 module github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar
 
 go 1.16
+
+require (
+	github.com/spf13/cobra v1.3.0
+	github.com/spf13/viper v1.10.1
+)
