+++ hw12_13_14_15_calendar/run/cmd/version.go
@@ -1,0 +1,38 @@
+package cmd
+
+import (
+	"encoding/json"
+	"fmt"
+	"os"
+
+	"github.com/spf13/cobra"
+)
+
+var (
+	release   = "UNKNOWN"
+	buildDate = "UNKNOWN"
+	gitHash   = "UNKNOWN"
+)
+
+func init() {
+	rootCmd.AddCommand(versionCmd)
+}
+
+var versionCmd = &cobra.Command{
+	Use:   "version",
+	Short: "Print the version number of calendar app",
+	Long:  `<Long version desc>`,
+	Run: func(cmd *cobra.Command, args []string) {
+		if err := json.NewEncoder(os.Stdout).Encode(struct {
+			Release   string
+			BuildDate string
+			GitHash   string
+		}{
+			Release:   release,
+			BuildDate: buildDate,
+			GitHash:   gitHash,
+		}); err != nil {
+			fmt.Printf("error while decode version info: %v\n", err)
+		}
+	},
+}
