+++ hw12_13_14_15_calendar/run/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar/run/cmd"
+)
+
+func main() {
+	cmd.Execute()
+}
